[Minor][C++] Forward declare joints in joint/fwd.hh
